chore: 戻るボタンの挙動を調整する
