+++ src/features/profile/layouts/profile-layout.tsx
@@ -1,16 +1,11 @@
-import { path } from "@/utils/path";
-import { ArrowBackIos } from "@mui/icons-material";
-import { Box, Stack } from "@mui/material";
+import { Box } from "@mui/material";
 import type { ReactNode } from "react";
-import { useNavigate } from "react-router-dom";
 
 type ProfileLayoutProps = {
   children: ReactNode;
 };
 
 export const ProfileLayout = ({ children }: ProfileLayoutProps) => {
-  const navigate = useNavigate();
-
   return (
     <Box
       sx={{
@@ -21,21 +16,6 @@ export const ProfileLayout = ({ children }: ProfileLayoutProps) => {
         paddingInline: "20px",
       }}
     >
-      <Stack
-        direction="row"
-        p={"10px"}
-        alignItems="center"
-        sx={{
-          width: "100%",
-        }}
-      >
-        <button
-          onClick={() => navigate(path.get().app.dashboard)}
-          type="button"
-        >
-          <ArrowBackIos sx={{ fontSize: "16px" }} />
-        </button>
-      </Stack>
       {children}
     </Box>
   );
